error with refactoring :3
