+++ portfolyo/__init__.py
@@ -1,13 +1,13 @@
 """Package to analyse and manipulate timeseries related to power and gas offtake portfolios."""
 
-from . import _version, dev, testing, tools
+from . import _version, dev, tools
 from .core import extendpandas  # extend functionalty of pandas
 from .core import suppresswarnings
 from .core.pfline import Kind, PfLine, Structure, create
 from .tools.freq import FREQUENCIES
 from .core.pfstate import PfState
-from .prices.hedge import hedge
-from .prices.utils import is_peak_hour
+from .tools.prices.hedge import hedge
+from .tools.prices.utils import is_peak_hour
 from .tools2.concat import general as concat
 from .tools2.intersect import indexable as intersection
 from .tools2.plot import plot_pfstates
@@ -18,6 +18,7 @@ from .tools.standardize import frame as standardize
 from .tools.tzone import force_agnostic, force_aware
 from .tools.unit import Q_, Unit, ureg
 from .tools.wavg import general as wavg
+from .tools import testing
 
 VOLUME = Kind.VOLUME
 PRICE = Kind.PRICE
