+++ portfolyo/core/pfline/flat_methods.py
@@ -1,10 +1,7 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
-
-from portfolyo import tools
-
-from ... import testing
+from ... import tools
 import pandas as pd
 
 if TYPE_CHECKING:
@@ -19,7 +16,7 @@ def __eq__(self: FlatPfLine, other: Any) -> bool:
     if not isinstance(other, self.__class__):
         return False
     try:
-        testing.assert_frame_equal(self.df, other.df, rtol=1e-7)
+        tools.testing.assert_frame_equal(self.df, other.df, rtol=1e-7)
         return True
     except AssertionError:
         return False
