--- portfolyo/testing/__init__.py
@@ -1,7 +1,0 @@
-from .testing import (
-    assert_frame_equal,
-    assert_index_equal,
-    assert_indices_compatible,
-    assert_series_equal,
-    assert_value_equal,
-)
