+++ portfolyo/tools/prices/hedge.py
@@ -4,7 +4,7 @@ from typing import Tuple
 
 import pandas as pd
 
-from .. import tools
+from ... import tools
 from . import convert, utils
