+++ portfolyo/tools/testing.py
@@ -6,8 +6,7 @@ from typing import Any
 import numpy as np
 import pandas as pd
 import pint
-from .. import tools
-from ..tools.unit import Q_
+from . import unit as tools_unit
 
 
 def assert_value_equal(left: Any, right: Any):
@@ -35,8 +34,8 @@ def assert_frame_equal(left: pd.DataFrame, right: pd.DataFrame, *args, **kwargs)
 
 @functools.wraps(pd.testing.assert_series_equal)
 def assert_series_equal(left: pd.Series, right: pd.Series, *args, **kwargs):
-    leftm, leftu = tools.unit.split_magn_unit(left)
-    rightm, rightu = tools.unit.split_magn_unit(right)
+    leftm, leftu = tools_unit.split_magn_unit(left)
+    rightm, rightu = tools_unit.split_magn_unit(right)
 
     # Magnitudes must be the same.
     leftm = leftm.replace([np.inf, -np.inf], np.nan)
@@ -65,21 +64,21 @@ def assert_indices_compatible(left: pd.DatetimeIndex, right: pd.DatetimeIndex):
 def assert_w_q_compatible(freq: str, w: pd.Series, q: pd.Series):
     """Assert if timeseries with power- and energy-values are consistent."""
     if freq == "15T":
-        assert_series_equal(q, w * Q_(0.25, "h"), check_names=False)
+        assert_series_equal(q, w * tools_unit.Q_(0.25, "h"), check_names=False)
     elif freq == "H":
-        assert_series_equal(q, w * Q_(1.0, "h"), check_names=False)
+        assert_series_equal(q, w * tools_unit.Q_(1.0, "h"), check_names=False)
     elif freq == "D":
-        assert (q >= w * Q_(22.99, "h")).all()
-        assert (q <= w * Q_(25.01, "h")).all()
+        assert (q >= w * tools_unit.Q_(22.99, "h")).all()
+        assert (q <= w * tools_unit.Q_(25.01, "h")).all()
     elif freq == "MS":
-        assert (q >= w * 27 * Q_(24.0, "h")).all()
-        assert (q <= w * 32 * Q_(24.0, "h")).all()
+        assert (q >= w * 27 * tools_unit.Q_(24.0, "h")).all()
+        assert (q <= w * 32 * tools_unit.Q_(24.0, "h")).all()
     elif freq == "QS":
-        assert (q >= w * 89 * Q_(24.0, "h")).all()
-        assert (q <= w * 93 * Q_(24.0, "h")).all()
+        assert (q >= w * 89 * tools_unit.Q_(24.0, "h")).all()
+        assert (q <= w * 93 * tools_unit.Q_(24.0, "h")).all()
     elif freq == "AS":
-        assert (q >= w * Q_(8759.9, "h")).all()
-        assert (q <= w * Q_(8784.1, "h")).all()
+        assert (q >= w * tools_unit.Q_(8759.9, "h")).all()
+        assert (q <= w * tools_unit.Q_(8784.1, "h")).all()
     else:
         raise ValueError(f"Uncaught value for freq: {freq}.")
