+++ tests/tools2/test_concat_error_cases.py
