+++ tests/utils.py
@@ -1,4 +1,5 @@
 """Create uniform ids for testcases."""
+
 from typing import Any, Dict
 
 import pandas as pd
